Do not repeat yourself
